+++ R/nflplotR-package.R
@@ -5,5 +5,6 @@
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
 #' @importFrom magrittr %>%
+#' @importFrom rlang .data
 ## usethis namespace: end
 NULL