+++ NEWS.md
@@ -8,3 +8,4 @@
 * Fixed an incompatible position argument in `scale_y_nfl()`. (v.0.0.9005)
 * Added the function `ggpreview()` which allows to preview a ggplot in it's actual dimensions. (v.0.0.9006)
 * `geom_nfl_logos()` now supports a `colour` aesthetic that colorizes the logos. (v0.0.9007)
+* Added the function `nfl_team_tiers()` that build an NFL team tiers ggplot, thanks to [Timo Riske](https://twitter.com/PFF_Moo) for the suggestion. (v.0.0.9008)